Move ProcessTestUtils and SparkSubmitTestUtils to core test sources
